+++ frontends/web/src/api/account.ts
@@ -398,6 +398,7 @@ export type TUTXO = {
   note: string;
   scriptType: ScriptType;
   addressReused: boolean;
+  isChange: boolean;
 };
 
 export const getUTXOs = (code: AccountCode): Promise<TUTXO[]> => {