+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Android: enable export logs feature
+- Label UTXOs that were created as change, as such, in coin control
 
 # 4.45.0
 - Bundle BitBox02 firmware version v9.21.0