+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Android: enable export logs feature
 - Label UTXOs that were created as change, as such, in coin control
 - Remove support for deprecated the Ethereum Goerli network
+- Revamp transaction history in account overview to improve legibility
 
 # 4.45.0
 - Bundle BitBox02 firmware version v9.21.0