+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Remove support for deprecated the Ethereum Goerli network
 - Revamp transaction history in account overview to improve legibility
 - Fix qrscanner when rotating the device or resizing the window
+- macOS: create a universal build that runs natively on arm64 and amd64
 
 # 4.45.0
 - Bundle BitBox02 firmware version v9.21.0